Add MindNode as a macOS app dependency
I'm getting more into mind mapping, if only for the spatially aware
tweak on top of my normal outlining routine. MindNode is pretty slick,
and works across both Mac and iOS, so I'm giving it a shot.

https://mindnode.com/

The most annoying part of all of this: MindNode is only available from
the Mac App Store :-/