Add script to copy message URL from Outlook.app
This builds on the technique from [this blog post][1] and adds both an
AppleScript and proxy application to copy and register an
'outlook://:messageId' URL.

Why would you do this? So I can process my Outlook.app inbox and queue
up messages that need further action (e.g., a reply or deep read) by
grabbing their URL and creating a task in my task manager of choice
(e.g., OmniFocus.app) that directs me to handle the follow-up action
with a clickable link back to the message. This keeps my email inbox a
place for triage, and my task manager as the place for action.

[1]: http://blog.hakanserce.com/post/outlook_automation_mac/